requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
torch>=2.1
tqdm>=4.66
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.5
python-dateutil>=2.9
